pyproject.toml
[project]
name = "security-utils"
version = "0.1.2"
description = "A collection of functions that are helpful for maintaining proper confidentiality."
readme = "README.md"
authors = [
    { name = "ktrin-u" }
]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
